+++ spec/rails_audit_log/graphql/queries/audit_log_entries_query_mixin_spec.rb
@@ -37,12 +37,12 @@ RSpec.describe RailsAuditLog::Graphql::Queries::AuditLogEntriesQueryMixin do
   describe "auditLogEntries field" do
     subject(:field) { fields.fetch("auditLogEntries") }
 
-    it "is a non-null list of non-null entries" do
-      expect(field.type.to_type_signature).to eq("[AuditLogEntry!]!")
+    it "returns a non-null connection type" do
+      expect(field.type.to_type_signature).to eq("AuditLogEntryConnection!")
     end
 
-    it "exposes exactly 6 arguments" do
-      expect(field.arguments.size).to eq(6)
+    it "exposes exactly 8 arguments (4 filters + 4 cursor pagination)" do
+      expect(field.arguments.size).to eq(8)
     end
 
     describe "filter arguments" do
@@ -63,19 +63,21 @@ RSpec.describe RailsAuditLog::Graphql::Queries::AuditLogEntriesQueryMixin do
       end
     end
 
-    describe "pagination arguments" do
-      it "has page argument defaulting to 1" do
-        arg = field.arguments.fetch("page")
-        expect(arg.type.non_null?).to be false
-        expect(arg.type.graphql_name).to eq("Int")
-        expect(arg.default_value).to eq(1)
+    describe "cursor pagination arguments" do
+      %w[first last].each do |arg_name|
+        it "has optional #{arg_name} Int argument" do
+          arg = field.arguments.fetch(arg_name)
+          expect(arg.type.non_null?).to be false
+          expect(arg.type.graphql_name).to eq("Int")
+        end
       end
 
-      it "has perPage argument defaulting to 25" do
-        arg = field.arguments.fetch("perPage")
-        expect(arg.type.non_null?).to be false
-        expect(arg.type.graphql_name).to eq("Int")
-        expect(arg.default_value).to eq(25)
+      %w[after before].each do |arg_name|
+        it "has optional #{arg_name} String argument" do
+          arg = field.arguments.fetch(arg_name)
+          expect(arg.type.non_null?).to be false
+          expect(arg.type.graphql_name).to eq("String")
+        end
       end
     end
   end
@@ -98,8 +100,6 @@ RSpec.describe RailsAuditLog::Graphql::Queries::AuditLogEntriesQueryMixin do
       allow(RailsAuditLog).to receive(:authenticate).and_return(nil)
       allow(RailsAuditLog::AuditLogEntry).to receive(:order).with(created_at: :desc).and_return(scope)
       allow(scope).to receive(:where).and_return(scope)
-      allow(scope).to receive(:limit).and_return(scope)
-      allow(scope).to receive(:offset).and_return(scope)
     end
 
     describe "#resolve_audit_log_entry" do
@@ -116,10 +116,8 @@ RSpec.describe RailsAuditLog::Graphql::Queries::AuditLogEntriesQueryMixin do
     end
 
     describe "#resolve_audit_log_entries" do
-      it "applies default pagination (page 1, per_page 25)" do
-        expect(scope).to receive(:limit).with(25).and_return(scope)
-        expect(scope).to receive(:offset).with(0).and_return(scope)
-        resolver.resolve_audit_log_entries
+      it "returns the ordered scope when no filters given" do
+        expect(resolver.resolve_audit_log_entries).to eq(scope)
       end
 
       it "filters by event" do
@@ -140,12 +138,6 @@ RSpec.describe RailsAuditLog::Graphql::Queries::AuditLogEntriesQueryMixin do
       it "filters by actor_id" do
         expect(scope).to receive(:where).with(actor_id: "7").and_return(scope)
         resolver.resolve_audit_log_entries(actor_id: "7")
-      end
-
-      it "calculates offset correctly for page 2" do
-        expect(scope).to receive(:limit).with(10).and_return(scope)
-        expect(scope).to receive(:offset).with(10).and_return(scope)
-        resolver.resolve_audit_log_entries(page: 2, per_page: 10)
       end
     end